Added some new files

+++ src/Testing_Phase/Mocks/MockSwap.sol
@@ -7,11 +7,11 @@ import "../../Interfaces/AggregatorV3Interface.sol";
 
 contract MockSwap {
     using SafeERC20 for IERC20;
-    
+
     IERC20 public usdc;
     IERC20 public usdy;
     AggregatorV3Interface internal priceFeed;
-    
+
     uint256 public constant USDC_DECIMALS = 6;
     uint256 public constant USDY_DECIMALS = 18;
 
@@ -26,10 +26,13 @@ contract MockSwap {
 
     function getLatestPrice() public view returns (uint256) {
         (
-            /* uint80 roundID */,
-            int price,
-            /*uint startedAt*/,
-            /*uint timeStamp*/,
+            /* uint80 roundID */
+            ,
+            int256 price,
+            /*uint startedAt*/
+            ,
+            /*uint timeStamp*/
+            ,
             /*uint80 answeredInRound*/
         ) = priceFeed.latestRoundData();
         require(price > 0, "Invalid price");
@@ -38,20 +41,20 @@ contract MockSwap {
 
     function swapUSDYtoUSDC(uint256 usdyAmount) external returns (uint256) {
         require(usdyAmount > 0, "Amount must be greater than 0");
-        
+
         // Get USDY price from sender
         usdy.safeTransferFrom(msg.sender, address(this), usdyAmount);
-        
+
         // Get latest USDC price from Chainlink and calculate USDC amount
         uint256 usdcPrice = getLatestPrice();
-        uint256 usdcAmount = (usdyAmount * usdcPrice * 10**USDC_DECIMALS) / (10**USDY_DECIMALS * 10**8); // Chainlink price feeds use 8 decimals
-        
+        uint256 usdcAmount = (usdyAmount * usdcPrice * 10 ** USDC_DECIMALS) / (10 ** USDY_DECIMALS * 10 ** 8); // Chainlink price feeds use 8 decimals
+
         // Check if we have enough USDC
         require(usdc.balanceOf(address(this)) >= usdcAmount, "Insufficient USDC liquidity");
-        
+
         // Transfer USDC to sender
         usdc.safeTransfer(msg.sender, usdcAmount);
-        
+
         return usdcAmount;
     }
 }
